% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox.R
\name{mapbox}
\alias{mapbox}
\alias{read_mvt}
\title{Mapbox Vector Tiles}
\usage{
read_mvt(x)
}
\arguments{
\item{x}{file path or raw vector with the mvt file}
}
\description{
Read Mapbox vector-tile (mvt) files and returns the list of layers.
}
