% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox.R
\name{mapbox}
\alias{mapbox}
\alias{read_mvt_data}
\alias{read_mvt_sf}
\title{Mapbox Vector Tiles}
\usage{
read_mvt_data(x)

read_mvt_sf(x)
}
\arguments{
\item{x}{file path or raw vector with the mvt file}
}
\description{
Read Mapbox vector-tile (mvt) files and returns the list of layers.
}
