% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox.R
\name{mapbox}
\alias{mapbox}
\alias{read_mvt_data}
\alias{read_mvt_sf}
\title{Mapbox Vector Tiles}
\usage{
read_mvt_data(data, zxy = NULL)

read_mvt_sf(data, zxy = NULL)
}
\arguments{
\item{data}{url, path or raw vector with the mvt data}

\item{zxy}{vector of length 3 with respsectively z (zoom), x (column) and y (row).
For file/url in the standard `../{z}/{x}/{y}.mvt` format, these are automatically
inferred from the input path.}
}
\description{
Read Mapbox vector-tile (mvt) files and returns the list of layers.
}
